+++ deno.json
@@ -1,9 +1,15 @@
 {
+  "name": "mastodon-to-bluesky",
+  "version": "1.1.0",
+  "license": "MIT",
   "tasks": {
     "dev": "deno run --watch main.ts"
   },
   "imports": {
-    "@atproto/api": "npm:@atproto/api@^0.15.5",
-    "@std/assert": "jsr:@std/assert@1"
+    "@atproto/api": "npm:@atproto/api@^0.15.27",
+    "@std/assert": "jsr:@std/assert@1",
+    "he": "npm:he@^1.2.0",
+    "axios": "npm:axios@^1.11.0",
+    "xmldom": "npm:xmldom@^0.6.0"
   }
 }
